chore: nats basic setup
