+++ src/services/collector/internal/models/nats.go
@@ -1,0 +1,5 @@
+package models
+
+type NatsConfig struct {
+	URL string `yaml:"url"`
+}
